+++ messageapp/Dockerfile
@@ -2,10 +2,10 @@ FROM node:latest
 
 EXPOSE 9001
 
-COPY exercise1/package*.json ./
+COPY package*.json ./
 
 RUN npm install
 
-COPY exercise1 ./
+COPY . ./
 
 CMD ["npm", "start"]
