+++ messageapp/app.js
@@ -1,0 +1,16 @@
+require("dotenv/config");
+
+require("./db");
+
+const express = require("express");
+
+const app = express();
+
+require("./config")(app);
+
+const allRoutes = require("./routes/index.routes");
+app.use("/api", allRoutes);
+
+require("./error-handling")(app);
+
+module.exports = app;
