+++ messageapp/server.js
@@ -1,6 +1,5 @@
 const app = require("./app");
 
-// ℹ️ Sets the PORT for our app to have access to it. If no env has been set, we hard code it to 3000
 const PORT = process.env.PORT || 9001;
 
 app.listen(PORT, () => {
