+++ messageapp/services/APIHandler.js
@@ -3,7 +3,7 @@ const axios = require('axios')
 module.exports = class ApiHandler {
     constructor() {
         this.axiosApp = axios.create({
-            baseURL: 'http://localhost:3000'
+            baseURL: 'http://messageapp:3000'
         })
     }
